Add the probability display
The line displays the probability of having seen the shiny after that
many encounters. Important to note that this does not affect the rate
at which shinies will happen. Even if the number says 95%, you are
still very unlikely to get a shiny next. See
https://en.wikipedia.org/wiki/Gambler%27s_fallacy

Why did I include this in the commit message.